Validators. Converter. Etc.

+++ MoneyManager/Model/Validators/DecimalValidator.cs
@@ -24,14 +24,14 @@ namespace MoneyManager.Model.Validators
 	{
 		public override ValidationResult Validate(object value, System.Globalization.CultureInfo cultureInfo)
 		{
-			Decimal amount;
+			float amount;
 			string temp = value as string;
 
 			if (string.IsNullOrWhiteSpace(temp))
 			{
 				return new ValidationResult(false, Properties.Resources.InputNotEmpty);
 			}
-			if (!Decimal.TryParse(temp, out amount))
+			if (!float.TryParse(temp, out amount))
 			{
 				return new ValidationResult(false, Properties.Resources.InputNotNumeric);
 			}
